Store paginated results.
Created helper method based on initial logic to make a GET request to the Instagram API that makes additional requests if there is a next_url attribute for the response object returned from the initial GET request and subsequent GET requests. Noticing a large load time for certain hashtags, i.e. it takes a few minutes for '#snow' and only a few seconds for '#pixleememories.' All results are stored however, checked using 'be rails c' and checking the results length for each search object.