Install Redis and add Sidekiq gem and Worker class.
Installed Redis in order to utilize Sidekiq to create a background job that would handle pagination on the server side while still being able to return the initial results of the first POST request. Steps taken include implementing a Worker class (InstagrabsWorker) and copying the helper functions from the application controller. If given time, will write a module to include the functions so that the application controller and the InstagrabsWorker class can both access the functions. Current "bug" identified is an 'undefined method [] for Nil:NilClass' that seems to point to the caption_empty? call on a result, but I am moving on for the sake of time and headaches and will try to tackle it later.